# doc/api.yaml
openapi: 3.0.0
paths:
...
  /track:
    get:
      operationId: TrackController_findMany
      summary: Get all tracks
      description: Gets all tracks
      parameters: []
      responses:
        "200":
          description: Track list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
        "401": &a5
          description: Access token is missing or invalid
      tags: &a6
        - Tracks
      security: &a7
        - access-token: []
    post:
      operationId: TrackController_create
      summary: Add new track
      description: Adds new track information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackDto"
      responses:
        "201":
          description: Added track information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: body does not contain required fields
        "401": *a5
      tags: *a6
      security: *a7
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: Get track by id
      description: Gets track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Track information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: id is invalid (not UUID)
        "401": *a5
        "404":
          description: Track not found
      tags: *a6
      security: *a7
    put:
      operationId: TrackController_update
      summary: Update track by id
      description: Updates track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: Updated track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: id is invalid (not UUID)
        "401": *a5
        "404":
          description: Track not found
      tags: *a6
      security: *a7
    delete:
      operationId: TrackController_delete
      summary: Delete track by id
      description: Deletes track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Track deleted
        "400":
          description: id is invalid (not UUID)
        "401": *a5
        "404":
          description: Track not found
      tags: *a6
      security: *a7
  /artist:
    get:
      operationId: ArtistController_findMany
      summary: Get all artists
      description: Gets all artists
      parameters: []
      responses:
        "200":
          description: Artist list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
        "401": &a8
          description: Access token is missing or invalid
      tags: &a9
        - Artists
      security: &a10
        - access-token: []
    post:
      operationId: ArtistController_create
      summary: Add new artist
      description: Adds new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: Added artist information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: body does not contain required fields
        "401": *a8
      tags: *a9
      security: *a10
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: Get artist by id
      description: Gets artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Artist information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: id is invalid (not UUID)
        "401": *a8
        "404":
          description: Artist not found
      tags: *a9
      security: *a10
    put:
      operationId: ArtistController_update
      summary: Update artist by id
      description: Updates artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: Updated artist information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: id is invalid (not UUID)
        "401": *a8
        "404":
          description: Artist not found
      tags: *a9
      security: *a10
    delete:
      operationId: ArtistController_delete
      summary: Delete artist by id
      description: Deletes artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Artist deleted
        "400":
          description: id is invalid (not UUID)
        "401": *a8
        "404":
          description: Artist not found
      tags: *a9
      security: *a10
  /album:
    get:
      operationId: AlbumController_findMany
      summary: Get all albums
      description: Gets all albums
      parameters: []
      responses:
        "200":
          description: Album list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
        "401": &a11
          description: Access token is missing or invalid
      tags: &a12
        - Albums
      security: &a13
        - access-token: []
    post:
      operationId: AlbumController_create
      summary: Add new album
      description: Adds new album
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: Added album information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: body does not contain required fields
        "401": *a11
      tags: *a12
      security: *a13
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: Get album by id
      description: Gets album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Album information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: id is invalid (not UUID)
        "401": *a11
        "404":
          description: Album not found
      tags: *a12
      security: *a13
    put:
      operationId: AlbumController_update
      summary: Update album by id
      description: Updates album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: Updated album information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: id is invalid (not UUID)
        "401": *a11
        "404":
          description: Album not found
      tags: *a12
      security: *a13
    delete:
      operationId: AlbumController_delete
      summary: Delete album by id
      description: Deletes album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Album deleted
        "400":
          description: id is invalid (not UUID)
        "401": *a11
        "404":
          description: Album not found
      tags: *a12
      security: *a13
...
  /favs/track/{id}:
    post:
      operationId: FavoritesTrackController_add
      summary: Add track to the favorites
      description: Adds track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "201":
          description: Track added to favorites
        "400":
          description: id is invalid (not UUID)
        "401": &a14
          description: Access token is missing or invalid
        "422":
          description: Track does not exist
      tags: &a15
        - Favorites
      security: &a16
        - access-token: []
    delete:
      operationId: FavoritesTrackController_delete
      summary: Delete track from favorites
      description: Deletes track from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Track deleted from favorites
        "400":
          description: id is invalid (not UUID)
        "401": *a14
        "404":
          description: Track not found
      tags: *a15
      security: *a16
  /favs/artist/{id}:
    post:
      operationId: FavoritesArtistController_add
      summary: Add artist to the favorites
      description: Adds artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "201":
          description: Artist added to favorites
        "400":
          description: id is invalid (not UUID)
        "401": &a17
          description: Access token is missing or invalid
        "422":
          description: Artist does not exist
      tags: &a18
        - Favorites
      security: &a19
        - access-token: []
    delete:
      operationId: FavoritesArtistController_delete
      summary: Delete artist from favorites
      description: Deletes artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Artist deleted from favorites
        "400":
          description: id is invalid (not UUID)
        "401": *a17
        "404":
          description: Artist not found
      tags: *a18
      security: *a19
  /favs/album/{id}:
    post:
      operationId: FavoritesAlbumController_add
      summary: Add album to the favorites
      description: Adds album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "201":
          description: Album added to favorites
        "400":
          description: id is invalid (not UUID)
        "401": &a20
          description: Access token is missing or invalid
        "422":
          description: Album does not exist
      tags: &a21
        - Favorites
      security: &a22
        - access-token: []
    delete:
      operationId: FavoritesAlbumController_delete
      summary: Delete album from favorites
      description: Deletes album from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Album deleted from favorites
        "400":
          description: id is invalid (not UUID)
        "401": *a20
        "404":
          description: Album not found
      tags: *a21
      security: *a22
info:
  title: Home Library Service API
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      in: header
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: aed621bd-fdb6-41ce-b092-d9861a2a2330
          description: Track id
        name:
          type: string
          description: Track name
          example: Courtesy Call
        artistId:
          type: string
          format: uuid
          example: 7d0a94e4-5703-4c99-b959-3688805d8610
          description: Artist ID
          nullable: true
        albumId:
          type: string
          format: uuid
          example: 4d8ee4c0-34f0-42e8-a1f7-c992163956fa
          description: Album ID
          nullable: true
        duration:
          type: integer
          example: 240
          description: Track duration in seconds
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Courtesy Call
        artistId:
          type: string
          format: uuid
          example: 7d0a94e4-5703-4c99-b959-3688805d8610
          description: Artist ID
          nullable: true
        albumId:
          type: string
          format: uuid
          example: 4d8ee4c0-34f0-42e8-a1f7-c992163956fa
          description: Album ID
          nullable: true
        duration:
          type: integer
          example: 240
          description: Track duration in seconds
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Courtesy Call
        artistId:
          type: string
          format: uuid
          example: 7d0a94e4-5703-4c99-b959-3688805d8610
          description: Artist ID
          nullable: true
        albumId:
          type: string
          format: uuid
          example: 4d8ee4c0-34f0-42e8-a1f7-c992163956fa
          description: Album ID
          nullable: true
        duration:
          type: integer
          example: 240
          description: Track duration in seconds
      required:
        - name
        - artistId
        - albumId
        - duration
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 5cdb3910-b777-4ce3-9eea-a9b62584541a
          description: Artist ID
        name:
          type: string
          example: Freddie Mercury
          description: Artist name
        grammy:
          type: boolean
          example: false
          description: Has Grammy award
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Freddie Mercury
          description: Artist name
        grammy:
          type: boolean
          example: false
          description: Has Grammy award
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Freddie Mercury
          description: Artist name
        grammy:
          type: boolean
          example: false
          description: Has Grammy award
      required:
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 316ceb21-4a71-4371-abf9-5f35c8cfef9c
          description: Album id
        artistId:
          type: string
          format: uuid
          description: Artist id
          example: 02bded4c-fe9a-41c2-807e-71d991c0c5db
          nullable: true
        name:
          type: string
          description: Album name
          example: Queen
        year:
          type: integer
          description: Year of release
          example: 1978
      required:
        - id
        - artistId
        - name
        - year
    CreateAlbumDto:
      type: object
      properties:
        artistId:
          type: string
          format: uuid
          description: Artist id
          example: 02bded4c-fe9a-41c2-807e-71d991c0c5db
          nullable: true
        name:
          type: string
          description: Album name
          example: Queen
        year:
          type: integer
          description: Year of release
          example: 1978
      required:
        - artistId
        - name
        - year
    UpdateAlbumDto:
      type: object
      properties:
        artistId:
          type: string
          format: uuid
          description: Artist id
          example: 02bded4c-fe9a-41c2-807e-71d991c0c5db
          nullable: true
        name:
          type: string
          description: Album name
          example: Queen
        year:
          type: integer
          description: Year of release
          example: 1978
      required:
        - artistId
        - name
        - year
...